rename class files to start with uppercase letter, move HealthBar class to classes directory
