+++ app/classes/EntityPrefab.js
