+++ app/classes/HealthBar.js
